Tuned hyperparams and switched to double precision

+++ src/lms/mod.rs
@@ -8,7 +8,7 @@ pub trait LM {
 
     fn reset(&mut self, dawg: &Dawg<usize>);
 
-    fn get_probability(&self, dawg: &Dawg<usize>, label: usize, good_turing: f32) -> f32;
+    fn get_probability(&self, dawg: &Dawg<usize>, label: usize, good_turing: f64) -> f64;
 
     fn update(&mut self, dawg: &Dawg<usize>, label: usize);
 }
